# kvserver: shard Raft scheduler