@erikgrinaker 2023-03-17
Going to run `kv0/enc=false/nodes=3/cpu=96`. Meanwhile, here's some mutex profiles from 50k unquiesced ranges on a 48-core machine:

```
$ go tool pprof -diff_base mutex-before.pb.gz mutex-after.pb.gz
(pprof) top -cum
Showing nodes accounting for -214.21s, 54.53% of 392.83s total
Dropped 269 nodes (cum <= 1.96s)
Showing top 10 nodes out of 23
      flat  flat%   sum%        cum   cum%
         0     0%     0%   -215.05s 54.74%  github.com/cockroachdb/cockroach/pkg/util/stop.(*Stopper).RunAsyncTaskEx.func2
  -214.21s 54.53% 54.53%   -214.21s 54.53%  sync.(*Mutex).Unlock
         0     0% 54.53%   -212.96s 54.21%  github.com/cockroachdb/cockroach/pkg/kv/kvserver.(*raftScheduler).worker
         0     0% 54.53%    173.44s 44.15%  github.com/cockroachdb/cockroach/pkg/kv/kvserver.(*raftScheduler).Start.func2
         0     0% 54.53%     39.20s  9.98%  github.com/cockroachdb/cockroach/pkg/kv/kvserver.(*Store).processTick
         0     0% 54.53%     38.29s  9.75%  github.com/cockroachdb/cockroach/pkg/util/hlc.(*Clock).NowAsClockTimestamp
         0     0% 54.53%     38.29s  9.75%  github.com/cockroachdb/cockroach/pkg/kv/kvserver.(*Replica).tick
         0     0% 54.53%    -18.72s  4.77%  sync.(*Cond).Wait
```

Before:

<img width="1446" alt="Screenshot 2023-03-17 at 14 23 27" src="https://user-images.githubusercontent.com/644420/225917206-6b54e988-a8be-4119-882f-671d2fc66524.png">

After:

<img width="1446" alt="Screenshot 2023-03-17 at 14 23 40" src="https://user-images.githubusercontent.com/644420/225917270-e342c7a3-b881-4fc5-822b-0f5d49b13f4a.png">