@erikgrinaker 2023-03-17
The results are pretty good (see PR description), as expected over in #96800.

Given the upside, I'm inclined to merge this before the branch cut. The changes are fairly well-contained, so the primary risk would be worker starvation with uneven shard load, but with 12 workers per shard the risk seems fairly minor. Let's have a look at the nightly benchmarks.