@erikgrinaker 2023-03-17
I added a commit with `raftSchedulerBatch` which avoids the O(shards) factor in `enqueueN` scans. However, on a 24-core cluster with 50k unquiesced ranges it didn't have any measurable impact on CPU usage, as one might expect. Unclear whether it's really an improvement, wdyt?