@erikgrinaker 2023-03-20
Looks like this is regressing slightly at lower core counts. Going to do another set with shard size 16 instead of 12.

```
name                          old ops/sec  new ops/sec  delta
kv0/enc=false/nodes=3/cpu=4    7.71k ± 5%   7.98k ± 2%     ~     (p=0.095 n=5+5)
kv0/enc=false/nodes=3/cpu=8    15.6k ± 3%   15.5k ± 1%     ~     (p=0.310 n=5+5)
kv0/enc=false/nodes=3/cpu=32   43.4k ± 2%   42.7k ± 5%     ~     (p=0.310 n=5+5)

name                          old p50      new p50      delta
kv0/enc=false/nodes=3/cpu=4     21.2 ± 6%    20.3 ± 3%     ~     (p=0.238 n=5+5)
kv0/enc=false/nodes=3/cpu=8     10.5 ± 0%    10.8 ± 3%     ~     (p=0.238 n=4+5)
kv0/enc=false/nodes=3/cpu=32    4.16 ± 1%    4.24 ± 8%     ~     (p=0.683 n=5+5)

name                          old p95      new p95      delta
kv0/enc=false/nodes=3/cpu=4     61.6 ± 5%    61.6 ± 2%     ~     (p=1.000 n=5+5)
kv0/enc=false/nodes=3/cpu=8     28.3 ± 4%    28.7 ± 2%     ~     (p=0.643 n=5+5)
kv0/enc=false/nodes=3/cpu=32    7.98 ± 2%    8.02 ± 1%     ~     (p=1.000 n=5+5)

name                          old p99      new p99      delta
kv0/enc=false/nodes=3/cpu=4     89.8 ± 7%    88.1 ± 0%     ~     (p=0.167 n=5+5)
kv0/enc=false/nodes=3/cpu=8     46.1 ± 0%    47.4 ± 3%     ~     (p=0.167 n=5+5)
kv0/enc=false/nodes=3/cpu=32    11.5 ± 0%    12.8 ± 6%  +11.30%  (p=0.008 n=5+5)
```