@pavelkalinnikov 2023-03-20
@erikgrinaker 
> Looks like this is regressing slightly at lower core counts. Going to do another set with shard size 16 instead of 12.

When you converge on some default shard/workers sizes, please add them to the unit test.